file updates
i removed the right click logic into a separate file to reduce the main files size
